server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-jose[cryptography]==3.3.0
cerebras.cloud.sdk
python-dotenv
